backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
slowapi==0.1.9
python-dotenv==1.0.0
alembic==1.13.1
itsdangerous==2.1.2
redis==5.0.1
cryptography==41.0.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.2
pyotp==2.9.0
qrcode[pil]==7.4.2